[refactor] Move caches package to another place
