+++ app/src/main/java/com/makentoshe/habrachan/model/posts/PostsResponseCache.kt
@@ -1,7 +1,7 @@
 package com.makentoshe.habrachan.model.posts
 
-import com.makentoshe.habrachan.common.model.cache.Cache
-import com.makentoshe.habrachan.common.model.cache.CacheStorage
+import com.makentoshe.habrachan.common.cache.Cache
+import com.makentoshe.habrachan.common.cache.CacheStorage
 import com.makentoshe.habrachan.common.model.network.postsalt.GetPostsRequest
 import com.makentoshe.habrachan.common.entity.posts.PostsResponse
