+++ app/src/main/java/com/makentoshe/habrachan/view/posts/PostsPageFragment.kt
@@ -11,7 +11,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.makentoshe.habrachan.R
-import com.makentoshe.habrachan.common.model.cache.Cache
+import com.makentoshe.habrachan.common.cache.Cache
 import com.makentoshe.habrachan.common.model.network.postsalt.GetPostsRequest
 import com.makentoshe.habrachan.common.model.network.postsalt.HabrPostsManager
 import com.makentoshe.habrachan.common.entity.posts.PostsResponse
